+++ idea/testData/multiFileIntentions/moveDeclarationToSeparateFile/moveClassToFile/after/source/test.kt
@@ -1,7 +1,5 @@
 package p
 
-import q.bar
-
 fun foo() {
 
 }