+++ idea/testData/multiFileIntentions/moveDeclarationToSeparateFile/optimizeImports/after/Foo.kt
@@ -1,0 +1,7 @@
+package foo.bar
+
+import baz.Baz
+
+class Foo {
+    val b = Baz()
+}