+++ idea/testData/multiFileIntentions/moveDeclarationToSeparateFile/optimizeImports/before/Baz.kt
@@ -1,0 +1,4 @@
+package baz
+
+class Baz {
+}