+++ idea/testData/multiFileIntentions/moveDeclarationToSeparateFile/optimizeImports/before/test.kt
@@ -1,0 +1,9 @@
+package foo.bar
+
+import baz.Baz
+
+class <caret>Foo {
+    val b = Baz()
+}
+
+class Bar